requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Babel==2.11.0
boto3==1.26.89
botocore==1.29.89
certifi==2022.12.7
charset-normalizer==3.0.1
dj-database-url==1.2.0
Django==4.1.6
django-ckeditor==6.5.1
django-countries==7.5.1
django-extensions==3.2.1
django-js-asset==2.0.0
django-on-heroku==1.1.2
django-phonenumber-field==7.0.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
Markdown==3.4.1
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.1.1
phonenumberslite==8.13.5
Pillow==9.4.0
protobuf==3.20.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.4.0
